src/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.15
distro==1.9.0
fastapi==0.104.1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.58.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
openai==1.59.7
packaging==24.2
propcache==0.2.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyGithub==2.5.0
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
sniffio==1.3.1
sse-starlette==1.8.2
starlette==0.27.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.24.0.post1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0